feature(login): split navigation into stack files and move screen styles/types into screen folders
