+++ src/navigation/AuthStackNavigator.tsx
@@ -1,0 +1,19 @@
+import React from 'react';
+import { createNativeStackNavigator, NativeStackScreenProps } from '@react-navigation/native-stack';
+import LoginScreen from '../features/auth/screens/login/LoginScreen';
+
+export type AuthStackParamList = {
+  Login: undefined;
+};
+
+const AuthStack = createNativeStackNavigator<AuthStackParamList>();
+
+export function AuthStackNavigator() {
+  return (
+    <AuthStack.Navigator screenOptions={{ headerShown: false }}>
+      <AuthStack.Screen name="Login" component={LoginScreen} />
+    </AuthStack.Navigator>
+  );
+}
+
+export type AuthStackScreenProps<T extends keyof AuthStackParamList> = NativeStackScreenProps<AuthStackParamList, T>;
