+++ src/navigation/HomeStackNavigator.tsx
@@ -1,0 +1,19 @@
+import React from 'react';
+import { createNativeStackNavigator, NativeStackScreenProps } from '@react-navigation/native-stack';
+import HomeScreen from '../features/home/screens/home/HomeScreen';
+
+export type HomeStackParamList = {
+  Home: undefined;
+};
+
+const HomeStack = createNativeStackNavigator<HomeStackParamList>();
+
+export function HomeStackNavigator() {
+  return (
+    <HomeStack.Navigator screenOptions={{ headerShown: false }}>
+      <HomeStack.Screen name="Home" component={HomeScreen} />
+    </HomeStack.Navigator>
+  );
+}
+
+export type HomeStackScreenProps<T extends keyof HomeStackParamList> = NativeStackScreenProps<HomeStackParamList, T>;
